Refactor CloudkitManager class
This refactor was done in order to make possible the choice
of which database to access. Unfortunately, zones are still
set to default.

For some unknown reason, some tests are broken due to a
'API violation - multiple calls made to -[XCTestExpectation fulfill]'
error which I haven't been able to solve.