Fix [CKRecord] to [StructModel] parser function
This commit make it so the parser function that parses from
CKRecord to StructModel always returns an array of StructModel.
If the array of CKRecord being parsed only contains
one CKRecord, the output of the function is an array of only
one StructModel.

Previously, this parser function would return a complete
StructModel with each CKReference as a StructModel which resulted
in a unnecessarily big fetch from CloudKit. In order to reduce
the size of each fetch, when a CKRecord is parsed into a
StructModel, its CKReferences are left nil. Performing the parsing
this way results in better performance in situations in which
CKReferences aren't in need of being shown to the use. I believe
this makes sense since the only scenario in which an user needs
to know, for instance, all the details of a Post is when it decides
to read the post. While in the Channel's posts screen all the
necessary information can be displayed without any information
from their respective CKReferences.