Add finished ChannelInteractor
Several new files were created to better work with Cloudkit and
the need to parse from CKRecords to StructModels. A new function
was created to do the opposite, transform from StructModels to
CKRecord.

Error handling still needs to be done. As of now, an error will
result in either nil or false being returned when a function
from ChannelInteractor fails.