Rewrite CoreDataManager to make use of async funcs
CoreDataManager has been rewritten to make use of asynchronous
functions. Some functions need to be rewritten to avoid too much
repetition.

This commit also changes every async functions in CloudKitManager
and all Interactors. It was written in such a way that most of the
completionHandlers would run on the main thread even though they
aren't related to UI.